Refactor ETL pipeline: shared type maps, pipeline runner, and batch inserts
- Add sql_sync/constants.py with unified SQLITE_TYPE_MAP and POSTGRES_TYPE_MAP,
  removing duplicate definitions from load.py and schema_excel_to_json.py
- Add sql_sync/run.py pipeline runner that wires extract -> transform -> load
  in a single command with --save-intermediates and --truncate flags
- Switch insert_table to executemany for batch inserts (one round-trip per table)
- Quote all column names in INSERT statements to handle reserved SQL keywords
- Remove unused --institution flag from extract.py (institution is set via DB_NAME)
- Skip __pycache__ and dunder dirs in schema_excel_to_json institution scanner

+++ tests/test_load.py
@@ -46,7 +46,7 @@ def mem_db():
 @pytest.mark.parametrize("data_type,expected", [
     ("bigint",    "INTEGER"),
     ("integer",   "INTEGER"),
-    ("numeric",   "REAL"),
+    ("numeric",   "NUMERIC"),
     ("string",    "TEXT"),
     ("text",      "TEXT"),
     ("date",      "TEXT"),
